api/requirements.txt
flask
flask_sqlalchemy
psycopg2-binary
python-dotenv
gunicorn
langgraph
langchain
langchain-google-genai
